Add files via upload
This is a project of student data visualization using python and its libraries that I made in my second year of college utilizing a data set which was made from scratch. The data file was just a sample to test out how this would work for university faculties to adopt since most data is stored in excel making it efficient to use.